# Swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "ToDo List",
        "description": "ToDo List Schnittstelle",
        "version": "0.0.1",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "petstore.swagger.io",
    "basePath": "/v2",
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
        "/list/{list_id}": {
            "get": {
                "summary": "Liefert alle Einträge einer Todo-Liste zurück",
                "description": "",
                "operationId": "uploadFile",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "list_id",
                        "in": "path",
                        "description": "ID of list",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/list"
                        }
                    },
                    "404": {
                        "description": "List not found",
                        "schema": {
                            "$ref": "#/definitions/list"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Löscht eine komplette Todo-Liste mit allen Einträgen.",
                "description": "",
                "operationId": "deletelist",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "list_id",
                        "in": "path",
                        "description": "List id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "List not found"
                    }
                }
            }
        },
        "/list": {
            "post": {
                "summary": "Fügt eine neue Todo-Liste hinzu.",
                "description": "",
                "operationId": "placeOrder",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "order placed for purchasing the pet",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/list"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/list"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/list/{list_id}/entry": {
            "post": {
                "summary": "Fügt einen Eintrag zu einer bestehenden Todo-Liste hinzu.",
                "description": "",
                "operationId": "postEntry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "list_id",
                        "in": "path",
                        "description": "List id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/entry"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/list/{list_id}/entry/{entry_id}": {
            "post": {
                "summary": "Aktualisiert einen bestehenden Eintrag.",
                "description": "",
                "operationId": "getentrybyuser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "order placed for purchasing the pet",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entry"
                        }
                    },
                    {
                        "name": "entry_id",
                        "in": "path",
                        "description": "List id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "list_id",
                        "in": "path",
                        "description": "List id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/entry"
                        }
                    },
                    "404": {
                        "description": "Eintrag Not Found"
                    }
                }
            },
            "delete": {
                "summary": "Löscht einen einzelnen Eintrag einer Todo-Liste.",
                "description": "",
                "operationId": "deleteentry",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "list_id",
                        "in": "path",
                        "description": "List id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "entry_id",
                        "in": "path",
                        "description": "List id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opperation Successful"
                    },
                    "404": {
                        "description": "Entry not found"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "summary": "Liefert eine Liste aller Benutzer zurück.",
                "description": "",
                "operationId": "getUser",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "summary": "Fügt einen neuen Benutzer hinzu.",
                "description": "",
                "operationId": "postUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "order placed for purchasing the pet",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    },
                    "404": {
                        "description": "User Not Found"
                    }
                }
            }
        },
        "/user/{user_id}": {
            "delete": {
                "summary": "Löscht einen Benutzer.",
                "description": "",
                "operationId": "deleteuser",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "USer id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opperation Successful",
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/user"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "user": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "username": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "entry": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "text": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "format": "int32"
                },
                "user": {
                    "$ref": "#/definitions/user"
                }
            }
        },
        "list": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entry"
                    }
                }
            }
        }
    }
}